Add bounds checks to backends
When attempting to open a corrupt file, read_xref_table_and_trailer may
try to read past the end of the file. This results in a Range with its
end before its start, and the library panics when slicing the underlying
data. This change checks for such cases, and returns an error instead of
panicking.