fix panic on malformed CIDToGID tables
If the number of bytes in the stream is odd, indexing c[1] will result in an
out of bounds access and will cause a panic.

This should not happen normally, since the spec says that this field should be
a list of u16, but some malformed PDF trigger this bug.