fix panic on malformed trailer ID
Found this bug by fuzzing.

If the trailer ID is an empty array, load_storage_and_trailer_password() will
panic when trying to access the first byte of an encrypted document.

Note that shouldn't happen on valid documents, the ID is supposed to be a
an array of two byte strings.